refactor: Refatora o BooksController para incluir o ResponseEntity nas rotas e customizar a resposta das requisicoes do cliente

+++ src/main/java/dev/cesar/LibrarySystem/Books/BooksController.java
@@ -1,5 +1,7 @@
 package dev.cesar.LibrarySystem.Books;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
@@ -14,28 +16,50 @@ public class BooksController {
     }
 
     @PostMapping("/create")
-    public BooksDTO createBook(@RequestBody BooksDTO book) {
-        return booksService.createBook(book);
+    public ResponseEntity<String > createBook(@RequestBody BooksDTO book) {
+        BooksDTO newBook = booksService.createBook(book);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Book successfully created: " + newBook.getTitle() + ". ID: " + newBook.getId());
     }
 
     @GetMapping("/list")
-    public List<BooksDTO> listAllBooks() {
-        return booksService.listAllBooks();
+    public ResponseEntity<List<BooksDTO>> listAllBooks() {
+        List<BooksDTO> books = booksService.listAllBooks();
+        return ResponseEntity.ok(books);
     }
 
     @GetMapping("/list/{id}")
-    public BooksDTO listBooksById(@PathVariable Long id) {
-        return booksService.listBookById(id);
+    public ResponseEntity<?> listBooksById(@PathVariable Long id) {
+        BooksDTO book = booksService.listBookById(id);
+        if (book != null) {
+            return ResponseEntity.status(HttpStatus.FOUND)
+                    .body(book);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Book of ID: " + id + " not found.");
+        }
     }
 
     @PutMapping("/update/{id}")
-    public BooksDTO updateBookById(@PathVariable Long id, @RequestBody BooksDTO book) {
-        return booksService.updateBookById(id, book);
+    public ResponseEntity<String> updateBookById(@PathVariable Long id, @RequestBody BooksDTO book) {
+        BooksDTO updatedBook = booksService.updateBookById(id, book);
+        if (updatedBook != null) {
+            return ResponseEntity.ok("Book of ID: " + id + " updated.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Book of ID: " + id + " not found.");
+        }
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteBookById(@PathVariable Long id) {
-        booksService.deleteBookById(id);
+    public ResponseEntity<String> deleteBookById(@PathVariable Long id) {
+        if (booksService.listBookById(id) != null) {
+            booksService.deleteBookById(id);
+            return ResponseEntity.ok("Book of ID: " + id + " deleted.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Book of ID: " + id + " not found.");
+        }
     }
